Support merging JSON over multiple browsers
Implementing creating merged JSON probe data over multiple browsers.
This will allow for easier processing and comparison of cross-browser
numbers.

- Enabled for JetStream, Speedometer and performance.mark probes
- Rename `RunGroup.csv_header` to `info` and make it return a dict

Drive-by-fixes:
- Change info keys to lower case
- Fix some typing annotations
- Update performance_entries.py to use direct imports
- Skip loading test_parse_example_page_config_file for google3 if
  hjson is not available

Reviewed-on: https://chromium-review.googlesource.com/c/crossbench/+/4089964
Commit-Queue: Camillo Bruni <cbruni@chromium.org>

+++ crossbench/benchmarks/loading.py
@@ -110,8 +110,8 @@ class InteractivePage(Page):
     result["actions"] = list(action.details_json() for action in self._actions)
     return result
 
-  def _get_duration(self) -> int:
-    duration: int = 0
+  def _get_duration(self) -> float:
+    duration: float = 0
     for action in self._actions:
       if action.duration is not None:
         duration += action.duration
